+++ pkg/barback/test/transformer_test.dart
@@ -7,6 +7,7 @@ library barback.test.transformer_test;
 import 'dart:async';
 
 import 'package:barback/barback.dart';
+import 'package:barback/src/utils.dart';
 import 'package:unittest/unittest.dart';
 
 import 'utils.dart';
@@ -17,25 +18,20 @@ main() {
   group("isPrimary", () {
     test("defaults to allowedExtensions", () {
       var transformer = new ExtensionTransformer(".txt .bin");
-      expect(transformer.isPrimary(new AssetId("pkg", "foo.txt")),
-          completion(isTrue));
+      expect(transformer.isPrimary(new AssetId("pkg", "foo.txt")), isTrue);
 
-      expect(transformer.isPrimary(new AssetId("pkg", "foo.bin")),
-          completion(isTrue));
+      expect(transformer.isPrimary(new AssetId("pkg", "foo.bin")), isTrue);
 
-      expect(transformer.isPrimary(new AssetId("pkg", "foo.nottxt")),
-          completion(isFalse));
+      expect(transformer.isPrimary(new AssetId("pkg", "foo.nottxt")), isFalse);
     });
 
     test("supports multi-level extensions with allowedExtensions", () {
       var transformer = new ExtensionTransformer(".dart.js");
-      expect(transformer.isPrimary(new AssetId("pkg", "foo.dart.js")),
-          completion(isTrue));
+      expect(transformer.isPrimary(new AssetId("pkg", "foo.dart.js")), isTrue);
 
-      expect(transformer.isPrimary(new AssetId("pkg", "foo.js")),
-          completion(isFalse));
-      expect(transformer.isPrimary(new AssetId("pkg", "foo.dart")),
-          completion(isFalse));
+      expect(transformer.isPrimary(new AssetId("pkg", "foo.js")), isFalse);
+
+      expect(transformer.isPrimary(new AssetId("pkg", "foo.dart")), isFalse);
     });
 
     test("throws an error for extensions without periods", () {
@@ -44,14 +40,11 @@ main() {
 
     test("allows all files if allowedExtensions is not overridden", () {
       var transformer = new MockTransformer();
-      expect(transformer.isPrimary(new AssetId("pkg", "foo.txt")),
-          completion(isTrue));
+      expect(transformer.isPrimary(new AssetId("pkg", "foo.txt")), isTrue);
 
-      expect(transformer.isPrimary(new AssetId("pkg", "foo.bin")),
-          completion(isTrue));
+      expect(transformer.isPrimary(new AssetId("pkg", "foo.bin")), isTrue);
 
-      expect(transformer.isPrimary(new AssetId("pkg", "anything")),
-          completion(isTrue));
+      expect(transformer.isPrimary(new AssetId("pkg", "anything")), isTrue);
     });
   });
 }